Fix for incompatible vermagic
Happens when building emgdbl for a different kernel, than the one which is currently used. Eg. when installing emgdbl, while you are doing a kernel update, it will build emgdbl against the old kernel and install it to the /lib/modules/$(uname -r)/ of the new kernel. After reboot into the new kernel the driver does not fit to the new kernel of course.